requirements.txt
PySide6>=6.6
httpx>=0.24
cryptography>=40.0
pycryptodome>=3.15
requests>=2.31
Pillow>=10.0
